backend/requirements.txt
Django>=3.2.3,<3.3
psycopg2>=2.8.6,<2.9
uWSGI>=2.0.19.1,<2.1
graphene-django==2.14.0
django-cors-headers==3.6.0
djangorestframework==3.12.4
